全局filter注册为全局方法，各全局方法合并到一个文件夹
