+++ src/common/js/commonMethods.js
